requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
beautifulsoup4==4.12.3
certifi==2023.5.7
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dateparser==1.1.8
distlib==0.3.7
Django==5.1.7
django-cors-headers==4.7.0
et_xmlfile==2.0.0
filelock==3.12.2
fonttools==4.56.0
frozendict==2.4.6
frozenlist==1.4.0
googletrans==4.0.0rc1
gTTS==2.5.4
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
h5py==3.10.0
hpack==3.0.0
hstspreload==2024.12.1
html5lib==1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
kiwisolver==1.4.8
logging==0.4.9.6
lxml==5.3.0
markdown-it-py==3.0.0
matplotlib==3.10.0
mdurl==0.1.2
meshio==5.3.5
multidict==6.0.4
multitasking==0.0.11
numpy==1.26.4
openpyxl==3.1.5
packaging==24.2
pandas==2.0.3
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.5
peewee==3.17.7
pillow==11.1.0
platformdirs==3.9.1
pycparser==2.22
pycryptodome==3.18.0
Pygments==2.19.1
pynite==1.4.2
pyparsing==3.2.1
pypdfium2==4.30.1
pytesseract==0.3.13
python-binance==1.0.17
python-box==7.3.0
python-dateutil==2.8.2
pytz==2023.3
regex==2023.6.3
requests==2.31.0
rfc3986==1.5.0
rich==13.9.4
scipy==1.12.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
ta==0.11.0
translator==0.0.9
tzdata==2023.3
tzlocal==5.0.1
ujson==5.8.0
urllib3==2.0.4
virtualenv==20.24.1
webencodings==0.5.1
websockets==11.0.3
yarl==1.9.2
yfinance==0.2.48
djangorestframework-simplejwt
tzdata>=2022.1
